+++ src/main/java/nl/enjarai/shared_resources/common/SharedResourcesPreLaunch.java
@@ -3,9 +3,6 @@ package nl.enjarai.shared_resources.common;
 import com.llamalad7.mixinextras.MixinExtrasBootstrap;
 import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
-import net.fabricmc.loader.impl.FabricLoaderImpl;
-import net.minecraft.util.Identifier;
-import nl.enjarai.shared_resources.api.GameResource;
 import nl.enjarai.shared_resources.api.GameResourceHelper;
 import nl.enjarai.shared_resources.api.GameResourceRegistry;
 import nl.enjarai.shared_resources.api.SharedResourcesEntrypoint;
@@ -14,10 +11,7 @@ import nl.enjarai.shared_resources.common.registry.GameResources;
 import nl.enjarai.shared_resources.versioned.Versioned;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.nio.file.Path;
-import java.util.Iterator;
 
 import static nl.enjarai.shared_resources.common.config.SharedResourcesConfig.CONFIG;